Migrate components to TS
