--- src/components/cart-item/CartItem.jsx
@@ -1,16 +1,0 @@
-import { CartItemContainer, ItemDetails } from './CartItem.styles'
-
-const CartItem = ({ cartItem }) => {
-  const { name, imageUrl, price, quantity } = cartItem
-  return (
-    <CartItemContainer>
-      <img src={imageUrl} alt={name} />
-      <ItemDetails>
-        <span>{name}</span>
-        <span>{quantity} x ${price}</span>
-      </ItemDetails>
-    </CartItemContainer>
-  )
-}
-
-export default CartItem
