+++ src/components/cart-item/CartItem.tsx
@@ -1,0 +1,24 @@
+import { FC } from 'react';
+import { CartItem as TCartItem } from '../../store/cart/cart.types';
+import { CartItemContainer, ItemDetails } from './CartItem.styles';
+
+type CartItemProps = {
+  cartItem: TCartItem;
+};
+
+const CartItem: FC<CartItemProps> = ({ cartItem }) => {
+  const { name, imageUrl, price, quantity } = cartItem;
+  return (
+    <CartItemContainer>
+      <img src={imageUrl} alt={name} />
+      <ItemDetails>
+        <span>{name}</span>
+        <span>
+          {quantity} x ${price}
+        </span>
+      </ItemDetails>
+    </CartItemContainer>
+  );
+};
+
+export default CartItem;
