+++ src/components/category-preview/CategoryPreview.styles.tsx
