--- src/components/checkout-item/CheckoutItem.jsx
@@ -1,37 +1,0 @@
-import { useDispatch , useSelector} from 'react-redux'
-import { addItemToCart, clearItemFromCart, removeItemFromCart} from '../../store/cart/cart.action'
-import {selectCartItems} from '../../store/cart/cart.selector'
-import { Arrow, BaseSpan, CheckoutItemContainer, ImageContainer, Quantity, RemoveButton, Value } from './CheckoutItem.styles'
-
-const CheckoutItem = ({ cartItem }) => {
-    const { name, imageUrl, price, quantity } = cartItem
-    
-    const cartItems = useSelector(selectCartItems)
-
-    const clearItemHandler = () => dispatch(clearItemFromCart(cartItems,cartItem))
-    const addItemHandler = () => dispatch(addItemToCart(cartItems,cartItem))
-    const removeItemHandler = () => dispatch(removeItemFromCart(cartItems,cartItem))
-
-    const dispatch = useDispatch()
-    return (
-        <CheckoutItemContainer>
-            <ImageContainer>
-                <img src={imageUrl} alt={name} />
-            </ImageContainer>
-            <BaseSpan className='name'>{name}</BaseSpan>
-            <Quantity>
-                <Arrow onClick={removeItemHandler}>
-                    &#10094;
-                </Arrow>
-                <Value>{quantity}</Value>
-                <Arrow onClick={addItemHandler}>
-                    &#10095;
-                </Arrow>
-            </Quantity>
-            <BaseSpan className='price'>{price}</BaseSpan>
-            <RemoveButton onClick={clearItemHandler}>&#10005;</RemoveButton>
-        </CheckoutItemContainer>
-    )
-}
-
-export default CheckoutItem
