--- src/components/directory-item/DirectoryItem.jsx
@@ -1,21 +1,0 @@
-import { useNavigate } from 'react-router-dom'
-import { BackGroundImage, Body, DirectoryItemContainer } from './DirectoryItem.styles'
-
-const DirectoryItem = ({ category }) => {
-    const { imageUrl, title , route} = category
-    const navigate = useNavigate()
-
-    const onNavigateHandler = ()=> navigate(route)
-
-    return (
-        <DirectoryItemContainer onClick={onNavigateHandler}>
-            <BackGroundImage imageUrl={imageUrl} />
-            <Body>
-                <h2>{title}</h2>
-                <p>Shop Now</p>
-            </Body>
-        </DirectoryItemContainer>
-    )
-}
-
-export default DirectoryItem
