+++ src/components/directory-item/DirectoryItem.tsx
@@ -1,0 +1,32 @@
+import { useNavigate } from 'react-router-dom';
+import {
+  BackGroundImage,
+  Body,
+  DirectoryItemContainer,
+} from './DirectoryItem.styles';
+
+import { FC } from 'react';
+import { DirectoryCategory } from '../directory/Directory';
+
+type DirectoryItemProps = {
+  category: DirectoryCategory;
+};
+
+const DirectoryItem: FC<DirectoryItemProps> = ({ category }) => {
+  const { imageUrl, title, route } = category;
+  const navigate = useNavigate();
+
+  const onNavigateHandler = () => navigate(route);
+
+  return (
+    <DirectoryItemContainer onClick={onNavigateHandler}>
+      <BackGroundImage imageUrl={imageUrl} />
+      <Body>
+        <h2>{title}</h2>
+        <p>Shop Now</p>
+      </Body>
+    </DirectoryItemContainer>
+  );
+};
+
+export default DirectoryItem;
