+++ src/components/index.ts
@@ -1,0 +1,14 @@
+export { default as Button } from './button/Button';
+export { default as CartDropdown } from './cart-dropdown/CartDropdown';
+export { default as CartIcon } from './cart-icon/CartIcon';
+export { default as CartItem } from './cart-item/CartItem';
+export { default as CategoryPreview } from './category-preview/CategoryPreview';
+export { default as CheckoutItem } from './checkout-item/CheckoutItem';
+export { default as DirectoryItem } from './directory-item/DirectoryItem';
+export { default as Directory } from './directory/Directory';
+export { default as FormInput } from './form-input/FormInput';
+export { default as PaymentForm } from './payment-form/PaymentForm';
+export { default as ProductCard } from './product-card/ProductCard';
+export { default as SignInForm } from './sign-in-form/SignInForm';
+export { default as SignUpForm } from './sign-up-form/SignUpForm';
+export { default as Spinner } from './spinner/Spinner';
