--- src/components/payment-form/PaymentForm.jsx
@@ -1,63 +1,0 @@
-import { CardElement, useElements, useStripe } from '@stripe/react-stripe-js'
-import { useState } from 'react'
-import { useSelector } from 'react-redux'
-import { selectCartTotal } from '../../store/cart/cart.selector'
-import { selectCurrentUser } from '../../store/users/user.selector'
-import { BUTTON_TYPE_CLASSES } from '../button/Button'
-import { FormContainer, PaymentButton, PaymentFormContainer } from './PaymentForm.styles'
-
-const PaymentForm = () => {
-
-    const stripe = useStripe()
-    const elements = useElements()
-
-    const amount = useSelector(selectCartTotal)
-    const currentUser = useSelector(selectCurrentUser)
-
-    const [isProcessingPayment, setIsProcessingPayment] = useState(false)
-
-    const paymentHandler = async (e) => {
-        e.preventDefault();
-        if (!stripe || !elements) {
-            return
-        }
-        setIsProcessingPayment(true)
-        const response = await fetch('/.netlify/functions/create-payment-intent', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ amount: amount * 100 })
-        }).then((res) => res.json())
-        const { paymentIntent: { client_secret } } = response
-
-        const paymentResult = await stripe.confirmCardPayment(client_secret, {
-            payment_method: {
-                card: elements.getElement(CardElement),
-                billing_details: {
-                    name: currentUser ? currentUser.displayName : 'Guest'
-                }
-            }
-        })
-        setIsProcessingPayment(false)
-
-        if (paymentResult.error) {
-            alert(paymentResult.error)
-        } else {
-            if (paymentResult.paymentIntent.status === 'succeeded') {
-                alert('Payment successfull!')
-            }
-        }
-    }
-    return (
-        <PaymentFormContainer>
-            <FormContainer onSubmit={paymentHandler}>
-                <h2>Credit Card Payment: </h2>
-                <CardElement />
-                <PaymentButton loading={isProcessingPayment} buttonType={BUTTON_TYPE_CLASSES.inverted}>Pay Now</PaymentButton>
-            </FormContainer>
-        </PaymentFormContainer>
-    )
-}
-
-export default PaymentForm
