+++ src/components/product-card/ProductCard.tsx
@@ -1,0 +1,43 @@
+import { useDispatch, useSelector } from 'react-redux';
+import { addItemToCart } from '../../store/cart/cart.action';
+import { selectCartItems } from '../../store/cart/cart.selector';
+import Button, { BUTTON_TYPE_CLASSES } from '../button/Button';
+import {
+  Footer,
+  Name,
+  Price,
+  ProductCardContainer,
+} from './ProductCard.styles';
+
+import { FC } from 'react';
+import { CategoryItem } from '../../store/categories/category.types';
+
+type ProductCardProps = {
+  product: CategoryItem;
+};
+
+const ProductCard: FC<ProductCardProps> = ({ product }) => {
+  const { name, price, imageUrl } = product;
+  const dispatch = useDispatch();
+
+  const cartItems = useSelector(selectCartItems);
+  const addProductToCart = () => dispatch(addItemToCart(cartItems, product));
+
+  return (
+    <ProductCardContainer>
+      <img src={imageUrl} alt={name} />
+      <Footer>
+        <Name>{name}</Name>
+        <Price>{price}</Price>
+      </Footer>
+      <Button
+        buttonType={BUTTON_TYPE_CLASSES.inverted}
+        onClick={addProductToCart}
+      >
+        Add To Cart
+      </Button>
+    </ProductCardContainer>
+  );
+};
+
+export default ProductCard;
