--- src/components/spinner/Spinner.jsx
@@ -1,11 +1,0 @@
-import {SpinnerContainer,SpinnerOverlay} from './Spinner.styles'
-
-const Spinner = () => {
-  return (
-    <SpinnerOverlay>
-        <SpinnerContainer/>
-    </SpinnerOverlay>
-  )
-}
-
-export default Spinner
