+++ src/components/spinner/Spinner.tsx
@@ -1,0 +1,11 @@
+import { SpinnerContainer, SpinnerOverlay } from './Spinner.styles';
+
+const Spinner = () => {
+  return (
+    <SpinnerOverlay>
+      <SpinnerContainer />
+    </SpinnerOverlay>
+  );
+};
+
+export default Spinner;
