--- src/routes/checkout/Checkout.jsx
@@ -1,37 +1,0 @@
-import { useSelector } from 'react-redux'
-import {selectCartItems,selectCartTotal} from '../../store/cart/cart.selector'
-import { CheckoutItem , PaymentForm} from '../../components'
-import { CheckoutContainer, CheckoutHeader, HeaderBlock, Total } from './Checkout.styles'
-
-const Checkout = () => {
-    const cartItems = useSelector(selectCartItems)
-    const cartTotal = useSelector(selectCartTotal)
-    return (
-        <CheckoutContainer>
-            <CheckoutHeader>
-                <HeaderBlock>
-                    <span>Product</span>
-                </HeaderBlock>
-                <HeaderBlock>
-                    <span>Description</span>
-                </HeaderBlock>
-                <HeaderBlock>
-                    <span>Quantity</span>
-                </HeaderBlock>
-                <HeaderBlock>
-                    <span>Price</span>
-                </HeaderBlock>
-                <HeaderBlock>
-                    <span>Remove</span>
-                </HeaderBlock>
-            </CheckoutHeader>
-            {cartItems.map((cartItem) => (
-                <CheckoutItem key={cartItem.id} cartItem={cartItem} />
-            ))}
-            <Total>Total: ${cartTotal}</Total>
-            <PaymentForm/>
-        </CheckoutContainer>
-    )
-}
-
-export default Checkout
