--- src/routes/home/Home.jsx
@@ -1,12 +1,0 @@
-import { Directory } from '../../components'
-
-const Home = () => {
-
-  
-
-  return (
-    <Directory/>
-  );
-}
-
-export default Home;
