+++ src/routes/home/Home.tsx
@@ -1,0 +1,7 @@
+import { Directory } from '../../components';
+
+const Home = () => {
+  return <Directory />;
+};
+
+export default Home;
