--- src/routes/index.js
@@ -1,8 +1,0 @@
-export { default as Authentication } from "./authentication/Authentication"
-export { default as CategoriesPreview } from "./categories-preview/CategoriesPreview"
-export { default as Category } from "./category/Category"
-export { default as Checkout } from "./checkout/Checkout"
-export { default as Home } from "./home/Home"
-export { default as Navigation } from "./navigation/Navigation"
-export { default as Shop } from "./shop/Shop"
-
