+++ src/routes/shop/Shop.tsx
@@ -1,0 +1,22 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { Route, Routes } from 'react-router-dom';
+import { CategoriesPreview, Category } from '..';
+import { fetchCategoriesStart } from '../../store/categories/category.action';
+
+const Shop = () => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchCategoriesStart());
+  }, []);
+
+  return (
+    <Routes>
+      <Route index element={<CategoriesPreview />} />
+      <Route path=":category" element={<Category />} />
+    </Routes>
+  );
+};
+
+export default Shop;
